+++ Practica2/src/entity/operand/inmediate/inmediate.cc
@@ -11,6 +11,6 @@
 
 #include "inmediate.h"
 
-double InmediateOperand::getValue(const DataMemory& registers) {
+int InmediateOperand::getValue(const DataMemory& registers) {
   return index_;
 }
